Moved the setup to conf file
